refactor components and type definitions to align with the Mike's superior structure.
had to comment out a few things in cards.tsx as the definitions.tx update broke it.  please review this.
