+++ hancrafted_haven/src/app/ui/products/ProductCard.tsx
@@ -1,6 +1,6 @@
 'use client';
 //import React, { useState } from 'react';
-import { Product } from '@/app/types/productTypes';
+import { Product } from '@/app/lib/definitions';
 import Image from 'next/image'
 
 interface ProductsProps {
